Do not do own logging by default.
Since pollon is meant to be used as a reusable package it shouldn't log
anything but just return errors.

Just for debugging purposes we want to do emit some debug logs.

This patch removes the use of capnslog and instead accepts a Logger. Logger is
an interface that mimic the stdlib log functions (with only the Print methods).
A pollon user can call SetLogger providing a stdlib Logger or a similar logger.

By default it'll not log anything.